# allow wrk array to be passed by user