--- a/src/Dierckx.jl
+++ b/src/Dierckx.jl
@@ -257,12 +257,9 @@ function Spline1D(x::AbstractVector, y::AbstractVector,
     return Spline1D(t, c, k, _translate_bc(bc), fp[])
 end
 
-# TODO: deprecate this?
-_evaluate(t::Vector{Float64}, c::Vector{Float64}, k::Int, x; bc::Int=0) =
-    __evaluate(t, c, k, x, bc)
 
-function __evaluate(t::Vector{Float64}, c::Vector{Float64}, k::Int,
-                   x::Vector{Float64}, bc::Int=0)
+function _evaluate(t::Vector{Float64}, c::Vector{Float64}, k::Int,
+                   x::Vector{Float64}, bc::Int)
     bc in (0, 1, 2, 3) || error("bc = $bc not in (0, 1, 2, 3)")
     m = length(x)
     xin = convert(Vector{Float64}, x)
@@ -279,8 +276,8 @@ function __evaluate(t::Vector{Float64}, c::Vector{Float64}, k::Int,
     return y
 end
 
-function __evaluate(t::Vector{Float64}, c::Vector{Float64}, k::Int,
-                   x::Real, bc::Int=0)
+function _evaluate(t::Vector{Float64}, c::Vector{Float64}, k::Int,
+                   x::Real, bc::Int)
     bc in (0, 1, 2, 3) || error("bc = $bc not in (0, 1, 2, 3)")
     y = Ref{Float64}(0)
     ier = Ref{Int32}(0)
@@ -295,23 +292,18 @@ function __evaluate(t::Vector{Float64}, c::Vector{Float64}, k::Int,
     return y[]
 end
 
-function evaluate(spline::Spline1D, x::AbstractVector)
-    xin = convert(Vector{Float64}, x)
-    __evaluate(spline.t, spline.c, spline.k, xin, spline.bc)
-end
+
+evaluate(spline::Spline1D, x::AbstractVector) =
+    _evaluate(spline.t, spline.c, spline.k,
+              convert(Vector{Float64}, x), spline.bc)
+
 
 evaluate(spline::Spline1D, x::Real) =
-    __evaluate(spline.t, spline.c, spline.k, x, spline.bc)
+    _evaluate(spline.t, spline.c, spline.k, x, spline.bc)
 
 
-# TODO: deprecate this?
-_derivative(t::Vector{Float64}, c::Vector{Float64}, k::Int, x;
-            nu::Int=1, bc::Int=0) =
-    __derivative(t, c, k, x, nu, bc)
-
-
-function __derivative(t::Vector{Float64}, c::Vector{Float64}, k::Int,
-                     x::Vector{Float64}, nu::Int=1, bc::Int=0)
+function _derivative(t::Vector{Float64}, c::Vector{Float64}, k::Int,
+                     x::Vector{Float64}, nu::Int, bc::Int)
     (1 <= nu <= k) || error("order of derivative must be positive and less than or equal to spline order")
     m = length(x)
     n = length(t)
@@ -329,8 +321,8 @@ function __derivative(t::Vector{Float64}, c::Vector{Float64}, k::Int,
     return y
 end
 
-function __derivative(t::Vector{Float64}, c::Vector{Float64}, k::Int,
-                     x::Real, nu::Int=1, bc::Int=0)
+function _derivative(t::Vector{Float64}, c::Vector{Float64}, k::Int,
+                     x::Real, nu::Int, bc::Int)
     (1 <= nu <= k) || error("order of derivative must be positive and less than or equal to spline order")
     n = length(t)
     wrk = Vector{Float64}(n)
@@ -351,13 +343,13 @@ end
 #       or should it be integrated with evaluate, above?
 #       (problem with that: derivative doesn't accept bc="nearest")
 # TODO: should `nu` be `d`?
-function derivative(spline::Spline1D, x::AbstractVector, nu::Int=1)
-    xin = convert(Vector{Float64}, x)
-    __derivative(spline.t, spline.c, spline.k, xin, nu, spline.bc)
-end
+derivative(spline::Spline1D, x::AbstractVector, nu::Int=1) =
+    _derivative(spline.t, spline.c, spline.k,
+                convert(Vector{Float64}, x), nu, spline.bc)
+
 
 derivative(spline::Spline1D, x::Real, nu::Int=1) =
-    __derivative(spline.t, spline.c, spline.k, x, nu, spline.bc)
+    _derivative(spline.t, spline.c, spline.k, x, nu, spline.bc)
 
 
 # TODO: deprecate this?
@@ -843,36 +835,35 @@ function ParametricSpline(u::AbstractVector, x::AbstractMatrix,
 end
 
 _evaluate(t::Vector{Float64}, c::Matrix{Float64}, k::Int,
-          x::Vector{Float64}; bc::Int=0) =
-    mapslices(v -> _evaluate(t, v, k, x, bc=bc), c, [2])
+          x::Vector{Float64}, bc::Int) =
+    mapslices(v -> _evaluate(t, v, k, x, bc), c, [2])
 
 _evaluate(t::Vector{Float64}, c::Matrix{Float64}, k::Int,
-          x::Real; bc::Int=0) =
-    vec(mapslices(v -> _evaluate(t, v, k, x, bc=bc), c, [2]))
+          x::Real, bc::Int) =
+    vec(mapslices(v -> _evaluate(t, v, k, x, bc), c, [2]))
 
 function evaluate(spline::ParametricSpline, x::AbstractVector)
     xin = convert(Vector{Float64}, x)
-    _evaluate(spline.t, spline.c, spline.k, xin, bc=spline.bc)
+    _evaluate(spline.t, spline.c, spline.k, xin, spline.bc)
 end
 
 evaluate(spline::ParametricSpline, x::Real) =
-    _evaluate(spline.t, spline.c, spline.k, x, bc=spline.bc)
+    _evaluate(spline.t, spline.c, spline.k, x, spline.bc)
 
 _derivative(t::Vector{Float64}, c::Matrix{Float64}, k::Int,
-            x::Vector{Float64}; nu::Int=1, bc::Int=0) =
-    mapslices(v -> _derivative(t, v, k, x; nu=nu, bc=bc), c, [2])
+            x::Vector{Float64}, nu::Int, bc::Int) =
+    mapslices(v -> _derivative(t, v, k, x, nu, bc), c, [2])
 
 _derivative(t::Vector{Float64}, c::Matrix{Float64}, k::Int,
-            x::Real; nu::Int=1, bc::Int=0) =
-    vec(mapslices(v -> _derivative(t, v, k, x; nu=nu, bc=bc), c, [2]))
+            x::Real, nu::Int, bc::Int) =
+    vec(mapslices(v -> _derivative(t, v, k, x, nu, bc), c, [2]))
 
-function derivative(spline::ParametricSpline, x::AbstractVector; nu::Int=1)
-    xin = convert(Vector{Float64}, x)
-    _derivative(spline.t, spline.c, spline.k, xin; nu=nu, bc=spline.bc)
-end
+derivative(spline::ParametricSpline, x::AbstractVector; nu::Int=1) =
+    _derivative(spline.t, spline.c, spline.k,
+                convert(Vector{Float64}, x), nu, spline.bc)
 
 derivative(spline::ParametricSpline, x::Real; nu::Int=1) =
-    _derivative(spline.t, spline.c, spline.k, x; nu=nu, bc=spline.bc)
+    _derivative(spline.t, spline.c, spline.k, x, nu, spline.bc)
 
 _integrate(t::Vector{Float64}, c::Matrix{Float64}, k::Int,
            a::Real, b::Real) =