This is another small performance improvement. (I didn't spot this until after I had submitted the previous pull request - my apologies.) 

This does not change default behaviour, but it gives the user the opportunity to pass a `wrk` array into `derivative` so that it doesn't have to be allocated every time. This gives again an advantage when one does many evaluations with scalar arguments. Here are the timings:

```
# OLD TIMINGS 
# evaluate 
#   0.018588 seconds (500.00 k allocations: 15.259 MB, 11.57% gc time)
#   0.017436 seconds (500.00 k allocations: 15.259 MB, 7.48% gc time)
# first derivative
#   0.020274 seconds (600.00 k allocations: 33.569 MB, 6.01% gc time)
#   0.020318 seconds (600.00 k allocations: 33.569 MB, 11.01% gc time)
# second derivative
#   0.027993 seconds (700.00 k allocations: 42.725 MB, 11.26% gc time)
#   0.026323 seconds (700.00 k allocations: 42.725 MB, 10.23% gc time)

# NEW TIMINGS: REMOVE TYPE INSTABILITY
# evaluate 
#   0.009857 seconds (200.00 k allocations: 3.052 MB)
#   0.009378 seconds (200.00 k allocations: 3.052 MB)
# first derivative
#   0.012261 seconds (300.00 k allocations: 19.837 MB, 8.79% gc time)
#   0.012860 seconds (300.00 k allocations: 19.837 MB, 6.69% gc time)
# second derivative
#   0.019107 seconds (400.00 k allocations: 28.992 MB, 11.40% gc time)
#   0.018948 seconds (400.00 k allocations: 28.992 MB, 11.45% gc time)

# NEW TIMINGS: WORK ARRAY PASSED AS AN ARGUMENT 
# evaluate 
#   0.010958 seconds (200.00 k allocations: 3.052 MB)
#   0.009265 seconds (200.00 k allocations: 3.052 MB)
# first derivative
#   0.008837 seconds (200.00 k allocations: 3.052 MB)
#   0.009624 seconds (200.00 k allocations: 3.052 MB)
# second derivative
#   0.009792 seconds (200.00 k allocations: 3.052 MB)
#   0.009251 seconds (200.00 k allocations: 3.052 MB)
```