@cortner 2017-04-11
I would actually implement this differently, but since I didn't want to change default behaviour I thought I'd first make this a straightforward PR but suggest this here: `wrk` could be allocated while `Spline1D` is created and be part of the `Spline1D` type.