@kbarbary 2017-04-11
I'm in favor of the general idea here, but I think it might be better to do what you suggest: make `wrk` part of the `Spline1D` type. I'd want to look and see if all the things you can do with `Spline1D` require the same size of work array. Since the contents of `Spline1D` are in some sense "private", this would be simply an internal optimization, with no API change. It would increase the memory footprint by ~50%, but I think that's OK.

As is, this PR doesn't *really* allow users to pass the work array, because methods starting with `_` are considered private.