@cortner 2017-04-12
Agree with all you say. Would you like me to update the PR or do you prefer to do this yourself.