@cortner 2017-04-12
This updated PR adds `wrk` to the `Spline1D` and `ParametricSpline` types and also fixes the type instability in `derivative(::ParametricSpline, ...)` that I fixed in an earlier PR for `Spline1D`.